Free (#1)
* adding

* build on this branch too

* Merge oops

* rm another license tier...

* Remove manifest, metrics, and verification refs

* updating action plugin

+++ docker-build/run-build.sh
@@ -2,21 +2,8 @@
 
 set -e
 
-USER=${SUDO_USER:-$USER}
-GROUP=$(id -gn $USER)
-
-# Run containers for each architecture
-if [[ "$1" == "x86_64" || -z "$1" ]]; then
-    sudo rm -rf build/
-    docker run --rm -t -v ./:/source --platform linux/amd64 -e STEAMOS -e UA_API_SECRET -e UA_API_SECRET_INTENTIONALLY_EMPTY "xr-driver:amd64"
-    sudo chown -R $USER:$GROUP out/
-fi
-
-if [[ "$1" == "aarch64" || -z "$1"  ]]; then
-    sudo rm -rf build/
-    docker run --rm -t -v ./:/source --platform linux/arm64 -e UA_API_SECRET -e UA_API_SECRET_INTENTIONALLY_EMPTY "xr-driver:arm64"
-    sudo chown -R $USER:$GROUP out/
-fi
+if [ -d build ]; then sudo rm -rf build/; fi
+podman run --rm -t --mount="type=bind,source=./,destination=/source" --platform linux/amd64 "xr-driver:amd64"
 
 # build directory structure is all owned by root because of docker, delete it all now
 sudo rm -rf build/
